move everything to new hpp file
