+++ beacon/include/marlin/beacon/ClusterDiscoverer.hpp
@@ -5,32 +5,13 @@
 #include <marlin/asyncio/udp/UdpFiber.hpp>
 #include <marlin/core/fibers/VersioningFiber.hpp>
 #include <marlin/core/fabric/Fabric.hpp>
-#include <marlin/asyncio/tcp/TcpOutFiber.hpp>
-#include <marlin/core/fibers/DynamicFramingFiber.hpp>
-#include <marlin/core/fibers/SentinelFramingFiber.hpp>
-#include <marlin/core/fibers/SentinelBufferFiber.hpp>
-#include <marlin/core/fibers/LengthFramingFiber.hpp>
-#include <marlin/core/fibers/LengthBufferFiber.hpp>
-#include <uv.h>
+#include <marlin/beacon/Grapher.hpp>
 #include <map>
-#include <string>
+
 #include <sodium.h>
-#include <spdlog/fmt/fmt.h>
-#include <rapidjson/document.h>
 
 #include "Messages.hpp"
 #include "PeerInfo.hpp"
-
-
-using namespace marlin::core;
-using namespace marlin::asyncio;
-
-
-std::map< std::string, std::string> MData;
-
-
-template<typename X>
-using SentinelFramingFiberHelper = SentinelFramingFiber<X, '\n'>;
 
 uint8_t* hextoint(std::string s){
 	//hex strings starting with 0x//
@@ -40,142 +21,7 @@ uint8_t* hextoint(std::string s){
         res[(i-2)/2] = (uint8_t)strtol(tmp, NULL, 16); 
     }
     return res;
-} 
-
-struct Grapher {
-	using FiberType = Fabric<
-		Grapher&,
-		TcpOutFiber,
-		DynamicFramingFiberHelper<
-			FabricF<SentinelFramingFiberHelper, SentinelBufferFiber>::type,
-			FabricF<LengthFramingFiber, LengthBufferFiber>::type
-		>::type
-	>;
-
-	SocketAddress dst;
-	std::map< std::string, std::string> MData;
-
-	void query_cb() {
-		SPDLOG_DEBUG("Timer hit: {}", dst.to_string());
-		auto& fiber = *new FiberType(std::forward_as_tuple(
-			*this,
-			std::make_tuple(),
-			std::make_tuple(std::make_tuple(
-				std::make_tuple(),
-				std::make_tuple()
-			))
-		));
-		(void)fiber.i(*this).dial(dst);
-	}
-
-	template<typename... Args>
-	Grapher(Args&&...) {
-		//t.template start<Grapher, &Grapher::query_cb>(0, 5000);
-	}
-
-	size_t state = 0;
-	size_t length = 0;
-
-	int did_recv(auto&& fiber, Buffer&& buf, SocketAddress addr [[maybe_unused]]) {
-		SPDLOG_DEBUG("Grapher: Did recv: {} bytes from {}: {}", buf.size(), addr.to_string(), std::string((char*)buf.data(), buf.size()));
-
-		if(state == 0) {
-			// headers
-			if(buf.size() == 2) {
-				// empty
-				if(length == 0) {
-					// still do not have length
-					fiber.o(*this).close();
-					return -1;
-				}
-
-				// go to next phase
-				state = 1;
-				fiber.o(*this).template transform<1>(std::make_tuple(), std::make_tuple());
-				fiber.o(*this).reset(length);
-				return 0;
-			} else if(std::memcmp(buf.data(), u8"content-length: ", 16) == 0) {
-				// parse length
-				auto length_str = std::string((char*)buf.data()+16, buf.size() - 18);
-				length = std::stol(length_str);
-			}
-			fiber.o(*this).reset(1000);
-		} else {
-			// body
-			state = 0;
-			SPDLOG_INFO("{}", std::string((char*)buf.data(), buf.size()));
-			rapidjson::Document JData;
-			JData.Parse((char*)buf.data(), buf.size());
-			if(!JData.HasParseError()){
-				auto& clusters = JData["data"]["clusters"];
-				for(rapidjson::SizeType i = 0; i < clusters.Size(); i++){
-					auto& cluster = clusters[i];
-					MData[cluster["clientKey"].GetString()] = cluster["id"].GetString();
-					SPDLOG_INFO("{}:{}", cluster["clientKey"].GetString(), cluster["id"].GetString());
-				}
-			}else{
-				SPDLOG_INFO("Not Parsed");
-			}
-			fiber.o(*this).close();
-		}
-		return 0;
-	}
-
-	template<typename FiberType>
-	int did_dial(FiberType& fiber, SocketAddress addr [[maybe_unused]]) {
-		SPDLOG_DEBUG("Grapher: Did dial: {}", addr.to_string());
-		fiber.o(*this).reset(1000);
-		auto query = Buffer(259).write_unsafe(
-			0,
-			(uint8_t*)"POST /subgraphs/name/marlinprotocol/staking HTTP/1.0\r\n"
-			"Content-Type: application/json\r\n"
-			"Content-Length: 150\r\n"
-			"\r\n"
-			"{\"query\": \"query { clusters(first: 1000, where: {networkId: \\\"0xaaaebeba3810b1e6b70781f14b2d72c1cb89c0b2b320c43bb67ff79f562f5ff4\\\"}){clientKey, id}}\"}",
-			259
-		);
-		fiber.o(*this).send(*this, std::move(query));
-		return 0;
-	}
-
-	template<typename FiberType>
-	int did_send(FiberType&, Buffer&& buf [[maybe_unused]]) {
-		SPDLOG_DEBUG("Grapher: Did send: {} bytes", buf.size());
-		return 0;
-	}
-
-	int did_close(auto& fiber) {
-		SPDLOG_DEBUG("Did close");
-		delete &fiber;
-		return 0;
-	}
-};
-
-int query() {
-	uv_getaddrinfo_t req;
-	auto res = uv_getaddrinfo(uv_default_loop(), &req, [](uv_getaddrinfo_t*, int, addrinfo* res) {
-		if(res != nullptr) {
-			auto& addr = *reinterpret_cast<SocketAddress*>(res->ai_addr);
-
-			SPDLOG_INFO("DNS result: {}", addr.to_string());
-
-			auto& g = *new Grapher(MData);
-			g.dst = addr;
-			g.query_cb();
-			// g.dst = SocketAddress::loopback_ipv4(18000);
-		}
-
-		uv_freeaddrinfo(res);
-	}, "graph.marlin.pro", "http", nullptr);
-	if(res != 0) {
-		SPDLOG_ERROR("DNS lookup error: {}", res);
-		return -1;
-	}
-
-	return uv_run(uv_default_loop(), UV_RUN_DEFAULT);
 }
-
-static int res = query();
 
 namespace marlin {
 namespace beacon {
@@ -212,7 +58,7 @@ private:
 	using LISTCLUSTER2 = LISTCLUSTER2Wrapper<BaseMessageType>;
 
 	FiberType fiber;
-
+	Grapher grapher;
 
 	// Discovery protocol
 	void send_DISCPROTO(FiberType& fiber, core::SocketAddress addr);
@@ -235,7 +81,6 @@ private:
 	void heartbeat_timer_cb();
 	asyncio::Timer heartbeat_timer;
 
-	
 public:
 	// Transport delegate
 	int did_dial(FiberType& fiber, core::SocketAddress addr, size_t type = 0);
@@ -324,6 +169,7 @@ void CLUSTERDISCOVERER::did_recv_LISTPROTO(
 
 		auto cluster_client_key = cluster_map[beacon_map[addr].first].address;
 		auto hexstring = fmt::format("0x{:spn}", spdlog::to_hex(cluster_client_key.data(), cluster_client_key.data()+20));
+		auto &MData = grapher.MData;
 		auto found = MData.find(hexstring);
 		if(found == MData.end()){
 			SPDLOG_INFO(
@@ -468,6 +314,7 @@ void CLUSTERDISCOVERER::did_recv_LISTCLUSTER2(
 
 		if constexpr (has_new_cluster) {
 			auto hexstring = fmt::format("0x{:spn}", spdlog::to_hex(cluster_client_key.data(), cluster_client_key.data()+20));
+			auto &MData = grapher.MData;
 			auto found = MData.find(hexstring);
 			if(found == MData.end()){
 				SPDLOG_INFO(
@@ -482,8 +329,6 @@ void CLUSTERDISCOVERER::did_recv_LISTCLUSTER2(
 				delegate->new_cluster(cluster_addr, cluster_id);
 			}
 		}
-
-		
 
 		cluster_map[cluster_addr].last_seen = asyncio::EventLoop::now();
 		cluster_map[cluster_addr].address = cluster_client_key;
